+++ src/Field/Base/InputData/FormModelInputData.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Form\InputData\FormModel;
+namespace Yiisoft\Form\Field\Base\InputData;
 
 use InvalidArgumentException;
 use Yiisoft\Form\Exception\PropertyNotSupportNestedValuesException;
 use Yiisoft\Form\Exception\StaticObjectPropertyException;
 use Yiisoft\Form\Exception\UndefinedObjectPropertyException;
 use Yiisoft\Form\Exception\ValueNotFoundException;
-use Yiisoft\Form\Field\Base\InputDataInterface;
+use Yiisoft\Form\Field\Base\InputData\InputDataInterface;
 use Yiisoft\Form\FormModelInterface;
 use Yiisoft\Validator\Helper\RulesNormalizer;
