+++ src/Field/Part/Hint.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Form\Field\Part;
 
 use InvalidArgumentException;
 use Stringable;
-use Yiisoft\Form\Field\Base\InputDataTrait;
+use Yiisoft\Form\Field\Base\InputData\InputDataTrait;
 use Yiisoft\Form\ThemeContainer;
 use Yiisoft\Html\Html;
 use Yiisoft\Widget\Widget;
