+++ src/PureField.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Form\InputData\Pure;
+namespace Yiisoft\Form;
 
+use Yiisoft\Form\Field\Base\InputData\PureInputData;
 use Yiisoft\Form\Field\Button;
 use Yiisoft\Form\Field\ButtonGroup;
 use Yiisoft\Form\Field\Checkbox;
@@ -32,7 +33,7 @@ use Yiisoft\Form\Field\Text;
 use Yiisoft\Form\Field\Textarea;
 use Yiisoft\Form\Field\Url;
 
-class Field
+class PureField
 {
     final public static function button(?string $content = null, array $config = []): Button
     {
