+++ tests/Field/Base/DateTimeInputFieldTest.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Form\Tests\Field\Base;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\FormModelInputData;
+use Yiisoft\Form\Field\Base\InputData\FormModelInputData;
 use Yiisoft\Form\Tests\Support\Form\DateForm;
 use Yiisoft\Form\Tests\Support\StubDateTimeInputField;
 use Yiisoft\Form\ThemeContainer;
