+++ tests/Field/DateTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Form\Tests\Field;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\FormModelInputData;
+use Yiisoft\Form\Field\Base\InputData\FormModelInputData;
 use Yiisoft\Form\Field\Date;
 use Yiisoft\Form\Tests\Support\Form\DateForm;
 use Yiisoft\Form\ThemeContainer;
