+++ tests/Field/Part/LabelTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Form\Tests\Field\Part;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\FormModelInputData;
+use Yiisoft\Form\Field\Base\InputData\FormModelInputData;
 use Yiisoft\Form\Field\Part\Label;
 use Yiisoft\Form\Tests\Support\Form\LabelForm;
 use Yiisoft\Form\Tests\Support\StringableObject;
