+++ tests/Field/TextTest.php
@@ -6,8 +6,8 @@ namespace Yiisoft\Form\Tests\Field;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\FormModelInputData;
-use Yiisoft\Form\InputData\Pure\PureInputData;
+use Yiisoft\Form\Field\Base\InputData\FormModelInputData;
+use Yiisoft\Form\Field\Base\InputData\PureInputData;
 use Yiisoft\Form\Field\Text;
 use Yiisoft\Form\Tests\Support\Form\TextForm;
 use Yiisoft\Form\ThemeContainer;
