+++ tests/Field/TextareaTest.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Form\Tests\Field;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\FormModelInputData;
+use Yiisoft\Form\Field\Base\InputData\FormModelInputData;
 use Yiisoft\Form\Field\Textarea;
 use Yiisoft\Form\Tests\Support\Form\TextareaForm;
 use Yiisoft\Form\ThemeContainer;
