+++ tests/FieldTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Form\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\InputData\FormModel\Field;
+use Yiisoft\Form\Field;
 use Yiisoft\Form\Tests\Support\Form\CheckboxForm;
 use Yiisoft\Form\Tests\Support\Form\CheckboxListForm;
 use Yiisoft\Form\Tests\Support\Form\DateForm;
